Simplify reading a file's tags
Before, I was using separate loops to first create a map (to ensure that
keywords weren't duplicated in output), and then looping through the
keys to build the list of keywords. I can just wrap this into one, with
an 'ok' check.

Can also just directly take element 1 of match, as it's only ever going
to be [FULLMATCH, CAPTUREGROUP].