Replace HashSet with BTreeSet
BTreeSet has ordered keys, which gives a slight performance increase. It
also enables me to remove the 'mut' from printing tags, as I no longer
need to sort the returned keys.

This _may_ change in future if I add back in the --numeirc option, as
I'll have to change the print and sort functionality.